requirements.txt
Flask
firebase-admin
google-cloud-firestore
Flask-CORS
Pillow